fix(stats): aplica os 9 achados do code review do PR #361
Correção:
- LlmInsightsView: try/catch/finally em handleRegenerateBacklog — rejeição
  da action (queda de rede) não deixa mais o botão preso em "Regenerando…"
  sem feedback; toast.error no catch.
- GabaritoSection: key do gabarito passa de respondentName (colide em
  homônimos e no fallback "Anônimo") para o id da resposta; o id agora é
  retornado por fetchGabaritoForComment (já era buscado e descartado) e
  mapeado do responseSnapshot.

Reuso/consolidação:
- formatAnswer duplicado removido de comment-card-utils; GabaritoSection e
  CommentsSplitView usam o de @/lib/reviews/verdict-format (superset: trata
  objetos, que antes rendiam "[object Object]").
- CommentsSplitView: cópias privadas de TYPE_LABELS/TYPE_COLORS/
  formatVerdictLabel/verdictVariant substituídas pelos exports de
  comment-card-utils (os helpers de verdict eram subconjunto — verdicts
  nota/sugestao/exclusao/etc. agora renderizam formatados lá também,
  corrigindo o drift existente).
- compareSemverDesc removido de useLlmErrorFiltering; lib/compare-version
  ganha compareVersions (3-way, malformada ordena como {0,0,0}) + testes.

Simplificação/altitude:
- ReviewComment/ResponseSnapshotEntry movidos para comment-card-utils
  (desfaz o ciclo type-only pai↔filho); CommentCard re-exporta para os
  consumidores externos.
- SuggestionActions: props estreitas (suggestionId/suggestionStatus) no
  lugar do ReviewComment inteiro — some o non-null assertion.
- ErrorFiltersToolbar: 18 props viram { fields, filtering } com a nova
  interface LlmErrorFilteringControls; hook retorna sortedCount.
- ExclusionActions: autoFocus na Textarea substitui useEffect+ref de foco;
  entrada correspondente removida do override no-event-handler do
  doctor.config.json.

Validação: react-doctor 90/100 (14 diag, sem novos), tsc 0 erros,
vitest 866/866.
