+++ frontend/src/components/stats/SuggestionActions.tsx
@@ -6,10 +6,10 @@ import { Badge } from "@/components/ui/badge";
 import { Button } from "@/components/ui/button";
 import { resolveSchemaSuggestion } from "@/actions/suggestions";
 import { toast } from "sonner";
-import type { ReviewComment } from "./CommentCard";
 
 interface SuggestionActionsProps {
-  comment: ReviewComment;
+  suggestionId: string;
+  suggestionStatus?: "pending" | "approved" | "rejected";
   projectId: string;
   isPending: boolean;
   isCoordinator?: boolean;
@@ -20,7 +20,8 @@ interface SuggestionActionsProps {
    "Revisar" abre EditFieldDialog pré-preenchido (via onResolve);
    salvar lá aprova com os valores finais editados. */
 export function SuggestionActions({
-  comment,
+  suggestionId,
+  suggestionStatus,
   projectId,
   isPending,
   isCoordinator,
@@ -31,7 +32,7 @@ export function SuggestionActions({
 
   return (
     <div className="flex items-center gap-2">
-      {comment.suggestionStatus === "pending" && isCoordinator && (
+      {suggestionStatus === "pending" && isCoordinator && (
         <>
           <Button
             variant="default"
@@ -50,7 +51,7 @@ export function SuggestionActions({
             disabled={suggestionPending}
             onClick={() => {
               startSuggestionAction(async () => {
-                const result = await resolveSchemaSuggestion(comment.suggestionId!, projectId, "rejected");
+                const result = await resolveSchemaSuggestion(suggestionId, projectId, "rejected");
                 if (result.error) toast.error(result.error);
                 else { toast.success("Sugestão rejeitada"); refresh(); }
               });
@@ -60,10 +61,10 @@ export function SuggestionActions({
           </Button>
         </>
       )}
-      {comment.suggestionStatus === "approved" && (
+      {suggestionStatus === "approved" && (
         <Badge className="text-xs bg-green-500/10 text-green-700">Aprovada</Badge>
       )}
-      {comment.suggestionStatus === "rejected" && (
+      {suggestionStatus === "rejected" && (
         <Badge className="text-xs bg-red-500/10 text-red-700">Rejeitada</Badge>
       )}
     </div>
